comp9900/requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
certifi==2022.9.24
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
Django==4.1.1
djangorestframework==3.14.0
idna==3.4
importlib-metadata==5.0.0
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
msgpack==1.0.4
packaging==21.3
Pygments==2.13.0
PyMySQL==1.0.2
pyparsing==3.0.9
pytz==2022.4
redis==4.3.4
requests==2.28.1
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.12
wrapt==1.14.1
zipp==3.8.1
